+++ _inc/lib/admin-pages/class.jetpack-react-page.php
@@ -252,6 +252,7 @@ class Jetpack_React_Page extends Jetpack_Admin_Page {
 				'isPublic'           => '1' == get_option( 'blog_public' ), // phpcs:ignore WordPress.PHP.StrictComparisons.LooseComparison
 				'isInIdentityCrisis' => Jetpack::validate_sync_error_idc_option(),
 				'sandboxDomain'      => JETPACK__SANDBOX_DOMAIN,
+				'errors'             => apply_filters( 'react_connection_errors_initial_state', array() ),
 			),
 			'connectUrl'                  => false == $current_user_data['isConnected'] // phpcs:ignore WordPress.PHP.StrictComparisons.LooseComparison
 				? Jetpack::init()->build_connect_url( true, false, false )
